+++ presto/_exceptions.py
@@ -3,3 +3,7 @@
 
 class InvalidSettingsError(ValueError):
     """Exception raised for invalid settings."""
+
+
+class MoleculeParameterisationError(RuntimeError):
+    """Exception raised when one or more molecules cannot be parameterised."""